begin moving routes to controllers
