+++ package.json
@@ -18,6 +18,7 @@
   "homepage": "https://github.com/ebirving/do_something_express_part1#readme",
   "dependencies": {
     "body-parser": "^1.13.3",
-    "express": "^4.13.3"
+    "express": "^4.13.3",
+    "hbs": "^3.1.0"
   }
 }
